Move profile_viewer files to spectrum tool plugin
