+++ glue/plugins/tools/spectrum_tool/__init__.py
@@ -1,0 +1,1 @@
+from .spectrum_tool import *
