+++ glue/plugins/tools/spectrum_tool/profile_viewer.py
