+++ glue/plugins/tools/spectrum_tool/tests/test_profile_viewer.py
